ルートアクセス（/）で直接ダッシュボードが表示されるように変更
